refactor: CountLotto 클래스 명 수정
